// src/app/core/Interfaces/j-branches/IAddBranchBody.ts
export interface IAddBranchBody {
  en_branch_location: string;
  ar_branch_location: string;
  en_branch_city: string;
  ar_branch_city: string;
  en_branch_address: string;
  ar_branch_address: string;
  branch_phone_1: string;
  branch_phone_2: string;
  branch_phone_3: string;
  status: string;
}